% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeATE.R
\name{computeATE}
\alias{computeATE}
\title{Estimation of average treatment effect}
\usage{
computeATE(x, z, mod, data, distribution)
}
\arguments{
\item{x}{Treatment variable (character string) currently treated as binary variable.}

\item{z}{Covariate variable (character string).}

\item{mod}{A \code{"glm"} object with logarithmic link function.
a standard Poisson regression is fitted.}

\item{data}{A data frame.}

\item{distribution}{Distribution of the covariate.
Currently, normal distribution (i.e., \code{"normal"}),
uniform distribution (i.e., \code{"uniform"}), Poisson distribution (i.e., \code{"poisson"}),
chisquare distribution (i.e., \code{"chisquare"}), and negative binomial distribution (i.e., \code{"negbin"}) are possible.}
}
\value{
List of average treatment effect and standard error.
}
\description{
Get point and standard error estimate for the average treatment effect given a fitted regression model.
}
\examples{
## Currently, no examples.


}
