% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEffects.R
\name{countEffects}
\alias{countEffects}
\title{Average treatment effects for count outcomes}
\usage{
countEffects(
  y,
  x,
  k = NULL,
  z = NULL,
  data,
  method = "nb",
  distribution = "condNormal",
  control = "default",
  measurement = list(),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{y}{Dependent variable (character string). Has to be the name of a count variable (i.e., non-negative integer).}

\item{x}{Treatment variable (character string) currently treated as binary variable.}

\item{z}{Covariate variable (character string).}

\item{data}{A data frame.}

\item{method}{By default a negative binomial regression model is fitted (i.e., \code{"nb"}). Alternatively, with \code{"poisson"}
a standard Poisson regression is fitted.}

\item{distribution}{Distribution of the covariate. By default, normal distribution is assumed (i.e., \code{"normal"}).
Currently, a uniform distribution (i.e., \code{"uniform"}), a Poisson distribution (i.e., \code{"poisson"}),
a chisquare distribution (i.e., \code{"chisquare"}), and a negative binomial distribution (i.e., \code{"negbin"}) are possible.}
}
\value{
Object of class countEffects.
}
\description{
A general function to estimate average treatment effects for a count outcome.
}
\examples{
## Example with normally distributed covariate:
m1 <- countEffects(y="dv", x="treat", z="pre", data=example01,
                   method="poisson", distribution="normal")
summary(m1)

}
