% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{dNA}
\alias{dNA}
\title{A simulated ACTIVE dataset}
\format{
An object of class \code{data.frame} with 2363 rows and 8 columns.
}
\usage{
dNA
}
\description{
A dataset containing simulated versions of the ACTIVE data subset used in
Kiefer & Mayer (2021)
}
\details{
#' @format A data frame with 2363 rows and 8 variables:
\describe{
  \item{dv}{dependent variable}
  \item{treat}{treatment variable}
  \item{gender}{gender covariate}
  \item{pretest}{pretest covariate}
  \item{ind1}{first indicator of depression}
  \item{ind2}{second indicator of depression}
  \item{ind3}{third indicator of depression}
}
}
\keyword{datasets}
