% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00documentation.R
\docType{data}
\name{example01}
\alias{example01}
\title{A first example}
\format{
An object of class \code{data.frame} with 333 rows and 3 columns.
}
\usage{
example01
}
\description{
basic dataset
}
\keyword{datasets}
